WebStorage/tests/submissions/Infraware/storage_session_key.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Web Storage</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>storage_session_key</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<script>

		if (('sessionStorage' in window) && window.sessionStorage !== null){
			sessionStorage.clear();

			sessionStorage.setItem("name", "user1");
			sessionStorage.setItem("age", "20");
			sessionStorage.setItem("a", "1");
			sessionStorage.setItem("b", "2");

			try {
				var keys = ["name","age","a","b"];
				function checkMatch(index){
					for (var i = 0; i < keys.length ; i++){
						if(sessionStorage.key(index) == keys[i]){
							delete keys[i];
							return true;
						}
					}
					return false;
				}
					
				test(function()
				{
					assert_true(checkMatch(0) && checkMatch(1) && checkMatch(2) && checkMatch(3), "key method is failed.");
				}, "key method test")
				
			} catch (e) {
				test(function()
				{
					assert_unreached("key method is failed.");
				}, "key method test")
			}
		} else {
			test(function()
			{
				assert_unreached("sessionStorage is not supported.");
			}, "sessionStorage support test")
		}	
	</script>

 </body>
</html>
